use the same Cloudpickler class for all Pythons
